fix up tests. move zip to it's own dir
