--- services/app-api/src/s3/zip/index.ts
@@ -1,1 +1,0 @@
-export { generateDocumentZip } from './generateZip'
